chore(server): move tests methods
