+++ packages/twenty-server/src/engine/metadata-modules/flat-index-metadata/types/flat-index-metadata.ts
