+++ packages/twenty-server/src/engine/metadata-modules/flat-index-metadata/utils/compare-two-flat-index-metadata.util.ts
@@ -1,7 +1,7 @@
 import diff from 'microdiff';
 import { FromTo } from 'twenty-shared/types';
 
-import { FlatIndexMetadata } from 'src/engine/workspace-manager/workspace-migration-v2/types/flat-index-metadata';
+import { FlatIndexMetadata } from 'src/engine/metadata-modules/flat-index-metadata/types/flat-index-metadata';
 import { transformMetadataForComparison } from 'src/engine/workspace-manager/workspace-sync-metadata/comparators/utils/transform-metadata-for-comparison.util';
 
 const flatIndexMetadataPropertiesToCompare = [
