+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/compute-updated-object-metadata-deleted-created-updated-index-matrix.util.ts
@@ -1,10 +1,10 @@
 import { FromTo } from 'twenty-shared/types';
 
+import { FlatIndexMetadata } from 'src/engine/metadata-modules/flat-index-metadata/types/flat-index-metadata';
 import {
   FlatObjectMetadata,
   FlatObjectMetadataWithoutFields,
 } from 'src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata';
-import { FlatIndexMetadata } from 'src/engine/workspace-manager/workspace-migration-v2/types/flat-index-metadata';
 import {
   CustomDeletedCreatedUpdatedMatrix,
   deletedCreatedUpdatedMatrixDispatcher,
